"Fix bugs in DeleteMapping"

+++ src/main/java/com/project/library/service/AuthorService.java
@@ -1,6 +1,7 @@
 package com.project.library.service;
 
 import com.project.library.entity.Author;
+import com.project.library.exception.ResourceNotFoundException;
 import com.project.library.repository.AuthorRepository;
 import org.springframework.stereotype.Service;
 
@@ -13,11 +14,13 @@ public class AuthorService {
     private final AuthorRepository authorRepository;
 
     public AuthorService(AuthorRepository authorRepository) {
+
         this.authorRepository = authorRepository;
     }
 
-    // ADD ALL AUTHORS
+    // ADD NEW AUTHORS
     public Author saveAuthor(Author author) {
+
         return authorRepository.save(author);
     }
 
@@ -27,12 +30,17 @@ public class AuthorService {
     }
 
     // GET AUTHOR BY ID
-    public Optional<Author> getAuthorById(Long authorId) {
-        return authorRepository.findById(authorId);
+    public Author getAuthorById(Long authorId) {
+        return authorRepository.findById(authorId)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("Author not found"));
     }
 
     // DELETE AUTHOR BY ID
     public void deleteAuthor(Long authorId) {
+        if (!authorRepository.existsById(authorId)) {
+            throw new ResourceNotFoundException("Author not found");
+        }
         authorRepository.deleteById(authorId);
     }
 }
